[#3919] Changed for uint32_t to user-defined literal and also created cpp file for StdInt

+++ nes-runtime/src/Nautilus/Interface/DataTypes/List/List.cpp
@@ -51,7 +51,7 @@ Value<> List::ListValueIterator::operator*() { return list.read(currentIndex); }
 bool List::ListValueIterator::operator==(const ListValueIterator& other) const { return currentIndex == other.currentIndex; }
 
 List::ListValueIterator& List::ListValueIterator::operator++() {
-    currentIndex = currentIndex + (uint32_t) 1;
+    currentIndex = currentIndex + 1_u32;
     return *this;
 }
