+++ nes-runtime/tests/UnitTests/Execution/Expressions/Functions/MinExpressionTest.cpp
@@ -72,7 +72,7 @@ TEST_F(MinExpressionTest, evaluateMinExpressionInteger) {
     }
     // UInt32
     {
-        auto resultValue = expression.eval(Value<UInt32>((uint32_t) 17), Value<UInt32>((uint32_t) 4));
+        auto resultValue = expression.eval(Value<UInt32>(17_u32), Value<UInt32>(4_u32));
         ASSERT_EQ(resultValue, (float) 4);
         ASSERT_TRUE(resultValue->getTypeIdentifier()->isType<Double>());
     }
